Re-implementation of AnnotatedString
This is a major commit that results from a complete rewrite of the AnnotatedString class, with the goal of closing issue #50 --which has been opened more than three years ago (!).

The new version of AnnotatedString relies on the Petit Poucet lineage tracking library, and many of its methods directly leverage the string manipulation functions that it provides. Contrary to what #50 suggested, the new class has richer functionalities, such as multiline regex find/replace, and its modular design should hopefully make it more robust than its predecessor. Thus, operations that required a line-by-line handling in the previous version can now be taken care of with a single global regex pattern. Line extraction is still supported, but in the longer term rules that work in this way could (and should) be refactored.