Fix: AnnotatedString.getLineOf results in an infinite loop under certain input #224 (#225)
* Fix: AnnotatedString.getLineOf results in an infinite loop under certain input #224

There seems to be a few issues with the existing loop that can cause it
to loop indefinitely. One potential issue is that the condition of the while loop checks for pos < s.length() before pos < index. If index is less than pos, the loop will never terminate because the condition pos < index will always be false.

Another potential issue is that the loop increments pos using pos = next_pos + CRLF_S, but it only does this if next_pos is less than s.length() and next_pos is less than index. If next_pos is equal to index, the loop will never terminate because pos will never be incremented.

For an example line that causes this, see example.tex.

To fix: used newer Java features which cleans up logic of iterating lines in string, the function will only iterate to a maximum of the strings length.

* Rewrite getLineOf for Java 8 Compatibility

Java 11 uses streams for its internal implementation of String.lines, so for Java 8 compatibility a while loop will do. This also should not result in an infinite loop.