Performance improvement on AnnotatedString
The previous version went through the construction of an explanation tree for a very long string of functions in order to find the input character range corresponding to an input character range. The current version only keeps in memory a map between input and output ranges, which gets updated each time the string is modified by an operation, somehow in line with a suggestion from @vmartinv. In some way, this can be seen as a "flattened" version of the explanation tree that only keeps track of the relationship between both ends of the chain.

This is on the way of addressing issue #234.