Fix to handling of empty inputs
When one of the files to analyze would result in an empty document after
being de-texed, an exception would be thrown and interrupt the analysis
of all files. This has been modified so that the analysis continues and
a warning is displayed only if *all* files to analyze were empty.